version 4

+++ frontend/src/screens/SingupScreen.jsx
@@ -9,26 +9,31 @@ import { Store } from '../Store.jsx';
 import { toast } from "react-toastify";
 import { getError } from "../utils.js";
 
-const SigninScreen = () => {
+const SignupScreen = () => {
     const navigate = useNavigate();
     const { search } = useLocation();
     const redirectInUrl = new URLSearchParams(search).get('redirect')
     const redirect = redirectInUrl ? redirectInUrl : '/';
 
+    const [name, setName] = useState('')
     const [email, setEmail] = useState('')
     const [password, setPassword] = useState('')
+    const [confirmpassword, setConfirmpassword] = useState('')
 
     const { state, dispatch: ctxDispatch } = useContext(Store);
     const { userInfo } = state;
     const submitHandler = async (e) => {
         e.preventDefault();
-
+        if (password !== confirmpassword) {
+            toast.error('Password is not match');
+            return;
+        }
 
         try {
-            const { data } = await Axios.post('http://localhost:5000/api/users/signin', {
-
+            const { data } = await Axios.post('http://localhost:5000/api/users/signup', {
+                name,
                 email,
-                password,
+                password
             })
             ctxDispatch({ type: 'USER_SIGNIN', payload: data })
             localStorage.setItem('userInfo', JSON.stringify(data));
@@ -47,10 +52,14 @@ const SigninScreen = () => {
         <>
             <Container className="small-container">
                 <Helmet>
-                    <title>Sign In</title>
+                    <title>Sign up</title>
                 </Helmet>
-                <h1 className='my-3'>Sign In</h1>
+                <h1 className='my-3'>Sign Up</h1>
                 <Form onSubmit={submitHandler}>
+                    <Form.Group className='mb-3' controlId="name">
+                        <Form.Label>Name</Form.Label>
+                        <Form.Control type='text' required onChange={(e) => setName(e.target.value)} />
+                    </Form.Group>
                     <Form.Group className='mb-3' controlId="email">
                         <Form.Label>Email</Form.Label>
                         <Form.Control type='email' required onChange={(e) => setEmail(e.target.value)} />
@@ -59,12 +68,16 @@ const SigninScreen = () => {
                         <Form.Label>Password</Form.Label>
                         <Form.Control type='password' required onChange={(e) => setPassword(e.target.value)} />
                     </Form.Group>
+                    <Form.Group className='mb-3' controlId="password">
+                        <Form.Label>Confirm Password</Form.Label>
+                        <Form.Control type='password' required onChange={(e) => setConfirmpassword(e.target.value)} />
+                    </Form.Group>
                     <div className='mb-3'>
-                        <Button type='submit'>Sign In</Button>
+                        <Button type='submit'>Sign Up</Button>
                     </div>
                     <div className='mb-3'>
-                        New customer ?{" "}
-                        <Link to={`/signup?rendirect=${redirect}`}>Create new Account</Link>
+                        Already have an account?{" "}
+                        <Link to={`/signin?rendirect=${redirect}`}>Sign In</Link>
                     </div>
                 </Form>
             </Container>
@@ -73,4 +86,4 @@ const SigninScreen = () => {
     )
 }
 
-export default SigninScreen
+export default SignupScreen
